code reorganization (split extraction functions to standalone files)
